학원에서..ㅇㅅㅇ 집에서 성희꺼랑 현은언니꺼 업데이트하기...ㅇㅅㅇ

+++ A_majoinUS/src/main/java/controller/PHE/AlarmController.java
@@ -10,7 +10,9 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.ResponseBody;
 import org.springframework.web.servlet.ModelAndView;
 
 import dao.PHE;
@@ -67,19 +69,26 @@ public class AlarmController {
 		return mav;
 	}
 	
-	
-	@RequestMapping(value = "/accetpMember")
+	@RequestMapping(value = "/acceptMember", method=RequestMethod.POST)
 	// method = { RequestMethod.GET, RequestMethod.POST }
+	@ResponseBody
 	public String acceptMember(@ModelAttribute("projectApplyAlarm") AlarmDTO dto,
-			@RequestParam("pj_num") int pj_num, @RequestParam("sender") String sender, String a_type) 
-	{
-		dao.acceptMember(sender, pj_num);
-		dao.removeFromAlarm_apply(sender, pj_num);
-		System.out.println("sender : " + sender);
-		System.out.println("pj_num : " + pj_num);
-		System.out.println("a_type : " + a_type);
-		System.out.println("=====================");
-		return "redirect:/aus/projectAlarm";
+			@RequestParam("pj_num") int pj_num, @RequestParam("sender") String sender/*, String a_type*/) 
+	{	
+		int check = dao.searchMember(sender,pj_num);
+		int Maxmembercount = dao.project_Max_count(pj_num);
+		int membercount = dao.projectMember(pj_num);
+		if(check == 0) {
+			if(Maxmembercount>membercount) {
+				dao.acceptMember(sender, pj_num);
+				dao.removeFromAlarm_apply(sender, pj_num);
+				return "register";
+			}else {
+				return "Max";
+			}
+		}else {
+			return "already";
+		}
 		
 		// 나에게 참가 신청을 하는거니까, 
 		//메세지 보낸 사람 아이디+프로젝트 번호 가져와서 Alarm 에서 "내 아이디는 필요없고 " 보낸이와 프로젝트 번호 가져와서 삭제
@@ -87,14 +96,26 @@ public class AlarmController {
 	}
 	
 	@RequestMapping(value = "/proposalAccept")
+	@ResponseBody
 	public String proposalAccept(@ModelAttribute("projectApplyAlarm") AlarmDTO dto,
-			@RequestParam("pj_num") int pj_num, @RequestParam("receiver") String receiver) {
+			@RequestParam("pj_num") int pj_num,HttpSession session) {
+		String receiver = (String)session.getAttribute("id");
 		
-		dao.acceptMember(receiver, pj_num);
-		dao.removeFromAlarm_invitation(receiver, pj_num);
-		System.out.println("receiver : " + receiver);
-		System.out.println("pj_num : " + pj_num);
-		return "redirect:/aus/projectAlarm";
+		int check = dao.searchMember(receiver,pj_num);
+		int Maxmembercount = dao.project_Max_count(pj_num);
+		int membercount = dao.projectMember(pj_num);
+		
+		if(check == 0) {
+			if(Maxmembercount>membercount) {
+				dao.acceptMember(receiver, pj_num);
+				dao.removeFromAlarm_invitation(receiver, pj_num);
+				return "register";
+			}else {
+				return "Max";
+			}
+		}else {
+			return "already";
+		}
 		
 		//다른 사람이 나에게 프로젝트에 초대한거니까
 		//그 팀장이 나 말고도 다른 사람에게 초대장을 보냈을 수도 있기 때문에 sender 로 잡지 말고
